refactor: split integration tests by crate boundary
Move engine tests to koda-core/tests/, keep CLI tests in koda-cli/tests/:

koda-core/tests/:
- file_tools_test.rs (10 tests) — path safety, file CRUD
- new_tools_test.rs (13 tests) — glob, todo, tool naming
- perf_test.rs (6 tests) — DB, grep, markdown, SSE throughput
- capabilities_test.rs (2 tests) — capabilities.md freshness

koda-cli/tests/:
- cli_test.rs (6 tests) — binary subprocess invocation
- regression_test.rs (18 tests) — REPL dispatch, input, display

Key changes:
- Replaced local re-implementations with koda_core:: imports
  (safe_resolve_path, Database)
- Recovered 55 integration tests that were orphaned by workspace split
- Total: 347 tests (was 292 after Phase 1-3)

Clippy clean.
