+++ koda-core/Cargo.toml
@@ -73,4 +73,6 @@ tree-sitter-typescript = "0.23.2"
 
 [dev-dependencies]
 tempfile = "3"
-path-clean = "1"
+tokio = { version = "1", features = ["full", "test-util"] }
+regex = "1"
+ignore = "0.4"
